Render acre point data only when fetched
